feat: 유효성 검사 및 로그인 관련 타입 파일 분할
