+++ app/(routes)/login/loginType.ts
@@ -1,0 +1,11 @@
+interface ValuesType {
+  email: string;
+  password: string;
+}
+
+interface ValidationType {
+  email: { isValid: boolean; message: string };
+  password: { isValid: boolean; message: string };
+}
+
+export type { ValidationType, ValuesType };
